Create separate subpages for schedule add and view
